% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_dat.R
\docType{data}
\name{heart_dat}
\alias{heart_dat}
\title{Heart Disease Dataset}
\format{
A data frame with 50000 observations and 21 variables
\describe{
 \item{Age}{}
 \item{Gender}{}
 \item{Weight}{}
 \item{Height}{}
 \item{BMI}{}
 \item{Smoking}{}
 \item{Alcohol_Intake}{}
 \item{Physical_Activity}{}
 \item{Diet}{}
 \item{Stress_Level}{}
 \item{Hypertension}{}
 \item{Diabetes}{}
 \item{Hyperlipidemia}{}
 \item{Family_History}{}
 \item{Previous_Heart_Attack}{}
 \item{Systolic_BP}{}
 \item{Diastolic_BP}{}
 \item{Heart_Rate}{}
 \item{Blood_Sugar_Fasting}{}
 \item{Cholesterol_Total}{}
 \item{Heart_Disease}{}
}
}
\usage{
data("heart_dat")
}
\description{
A synthetic heart disease dataset included with the HeartR package
}
\keyword{datasets}
