% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_heart.R
\name{heart_data}
\alias{heart_data}
\title{Create a Heart Disease Data Object}
\usage{
heart_data(data, response)
}
\arguments{
\item{data}{A data.frame containing the heart disease dataset with predictors
and the response variable.}

\item{response}{A character string specifying the column name of the response
variable.}
}
\value{
An object of class `"heart_data"` containing:
  \itemize{
    \item `data`: the input dataset
    \item `response`: the name of the response variable
  }
}
\description{
`heart_data()` creates a structured object containing a dataset and the response
variable indicating heart disease. This object is used as input for modeling
and plotting functions.
}
\examples{
\dontrun{
heart <- heart_data(data = heart_disease_data, response = "Heart_Disease")
heart
}

}
