% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_heart.R
\name{model}
\alias{model}
\title{Fit a Heart Disease Model}
\usage{
model(x, method = c("glm", "rf"))
}
\arguments{
\item{x}{A `heart_data` object containing the dataset and response variable.}

\item{method}{Character. The modeling method to use:
\itemize{
  \item `"glm"`: Logistic regression
  \item `"rf"`: Random forest
}}
}
\value{
An object of class `"heart_model"` containing:
  \itemize{
    \item `fit`: the fitted model object
    \item `method`: modeling method used
    \item `data`: input dataset
    \item `response`: response variable name
  }
}
\description{
`model()` fits a predictive model to a `heart_data` object. Users can choose
between logistic regression (for binary classification, given the fact that
some of are in 0s and 1s) and random forest.
}
\examples{
\dontrun{
heart <- heart_data(data = heart_disease_data, response = "Heart_Disease")

# Logistic regression
heart_glm <- model(heart, method = "glm")

# Random forest
heart_rf <- model(heart, method = "rf")
}

}
