% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summ_heartr.R
\name{print.summ_heartr}
\alias{print.summ_heartr}
\title{Print summ_heartr output}
\usage{
\method{print}{summ_heartr}(
  x,
  bordercolour = "pink",
  fillcolour = "#FFEDFB",
  font = "Roboto Condensed",
  ...
)
}
\arguments{
\item{x}{An object of class \code{"summ_heartr"} outputted from the
summ_heartr function.}

\item{bordercolour}{Colour of the cell borders. Must be in quotations.
Default colour of pink is used.}

\item{fillcolour}{Colour of the Variable cells. Must be in quotations.
Default colour of #FFEDFB is used}

\item{font}{Google style used in the table. Must be a Google font which can
be found \href{https://fonts.google.com/}{here}.}

\item{...}{Catches unused arguments to \code{print} (not currently
implemented).}
}
\value{
An aesthetically pleasing table nicely displaying the summary
statistics.
}
\description{
Provides a nice customisable table for the output of the summ_heartr function.
}
\examples{
# example code




}
\seealso{
\code{\link{print}}, \code{\link{summ_heartr}}
}
\author{
Ciara Olohan - <\email{ciara.olohan.2023@mumail.ie}>
}
