% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulation}
\usage{
simulation(seed = 403, n = NULL)
}
\arguments{
\item{seed}{random seed}

\item{n}{number of rows to data}
}
\value{
A synthetic dataset with continuous, categorical and outcome variables
}
\description{
This function generates reproducible synthetic datasets
}
