% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_heartr.R
\name{summ_heartr}
\alias{summ_heartr}
\title{summ_heartr}
\usage{
summ_heartr(...)
}
\arguments{
\item{...}{One or more numeric variables.}
}
\value{
An object of class summ_heartr containing the following statistics
for each input variable:
mean, minimum, 1st quartile, median, 3rd quartile, maximum, standard
deviation and variance.
}
\description{
Creates a list of the following summary statistics for one or more numeric
variables:
mean, minimum, 1st quartile, median, 3rd quartile, maximum, standard
deviation and variance.
}
\note{
A dedicated \code{\link{print.summ_heartr}} method is provided for
objects of class \code{"summ_heartr"}.
}
\examples{
summ_heartr(heart_dat$age, heart_dat$cholesterol)



}
\seealso{
\code{\link{print.summ_heartr}}
}
\author{
Ciara Olohan - <\email{ciara.olohan.2023@mumail.ie}>
}
