% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_heartrplot.R
\name{summ_heartrplot}
\alias{summ_heartrplot}
\title{An interactive histogram plot with mean and median}
\usage{
summ_heartrplot(data, var)
}
\arguments{
\item{data}{A data frame containing the variable to be plotted.}

\item{var}{A character string that gives the name of the numeric variable to
plot}
}
\value{
An interactive histogram produced by \code{plotly::ggplotly()}
}
\description{
This function create a histogram based on the selected variable from the data
frame. The vertical lines shows the mean and the median, and the plot is
converted to an interactive plot using \pkg{plotly}.
}
\examples{
<<<<<<< HEAD:man/summ_heartrplot.Rd
summ_heartrplot(heart_dat, "age")
=======
plot_summary(heart_dat, "age")
>>>>>>> 977d938e1dd2cf7bc94b9aee14204f04e7ec7e27:man/plot_summary.Rd


}
\seealso{
\code{\link{summ_heartr}}, \code{\link{print.summ_heartr}}
}
