% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_heart}
\alias{plot_heart}
\title{An interactive histogram plot with mean and median}
\usage{
plot_heart(
  heart,
  type = c("hist", "density", "scatter", "boxplot", "correlation"),
  var_x = NULL,
  var_y = NULL
)
}
\arguments{
\item{data}{A data frame containing the variable to be plotted.}

\item{var}{A character string that gives the name of the numeric variable to
plot}
}
\value{
An interactive histogram produced by \code{plotly::ggplotly()}
}
\description{
This function create a histogram based on the selected variable from the data
frame. The vertical lines shows the mean and the median, and the plot is
converted to an interactive plot using \pkg{plotly}.
}
