% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{simulation}
\usage{
simulation(
  seed = 403,
  n = NULL,
  data = NULL,
  contVars = NULL,
  catVars = NULL,
  outcome = "target",
  thres = 0.5,
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{seed}{random seed for reproducibility}

\item{n}{number of rows in the synthetic datasets. Deaults to the number of rows in
\code{data}.}

\item{data}{dataset the user wishes to replicate. If \code{NULL}, heart
disease dataset is used.}

\item{contVars}{character vector of continuous numeric variables in \code{data}.
If \code{NULL} then all numeric variables are used.}

\item{catVars}{character vector of categorical variables in \code{data}. If
\code{NULL} then all factor/ categorical variables, excluding the outcome,
are used}

\item{outcome}{name of the binary outcome variable in \code{data}}

\item{thres}{numeric threshold between 0 and 1 used to convert predicted probabilities
into binary outcomes. Default set to 0.5.}

\item{mu}{optional named vector of means used to simulate continuous variables}

\item{sigma}{optional covariance matrix to simulate continuous variables.}
}
\value{
A synthetic dataset with continuous, categorical and outcome variables
}
\description{
This function generates reproducible synthetic datasets, suitable for binary
classification. The current version only works for datasets in which the response
variable has exactly two possible values.
}
\note{
Datasets like \code{mtcars} contains no categorical variables. However, variables
(e.g. \code{cyl}, \code{gear}) may be considered to be of that nature. Ensure
variables in any dataset is of the correct type before passing it to this function.
}
\examples{
\dontrun{
  sim_mtcars <- simulation(data = mtcars, outcome = "am")
}

sim_data <- simulation(seed = 123)

sim_data2 <- simulation(seed = 23, n = 50)
sim_data2

sim_data3 <- simulation(
  data = heart_dat,
  n = 100,
  mu = c(age = 55, cholesterol = 220, restingBP = 130)
)

sim_dat4 <- simulation(
  data = heart_dat,
  n = 80,
  contVars = c("age", "cholesterol","restingBP"),
  catVars = c("sex","chestPT","exerAngina")
)

sim_dat5 <- simulation(
  data = heart_dat,
  n = 100,
  contVars = c("age", "cholesterol","restingBP"),
  catVars = c("sex","chestPT","exerAngina"),
  sigma = cov(heart_dat[, c("age", "cholesterol","restingBP")])
 )

}
\author{
Faridat Adeniji - <\email{faridaadeniji@gmail.com}>
}
